fix: Reorganize, add tests
