+++ py/torch_tensorrt/__init__.py
@@ -93,6 +93,8 @@ from torch_tensorrt._Device import Device
 from torch_tensorrt._TRTModuleNext import TRTModuleNext
 
 from torch_tensorrt import fx
+from torch_tensorrt import dynamo
+from torch_tensorrt.dynamo import torch_compile
 
 
 def _register_with_torch():
