--- py/torch_tensorrt/dynamo/lowering/__init__.py
@@ -1,2 +1,0 @@
-from torch_tensorrt.dynamo.lowering._decompositions import get_decompositions
-from torch_tensorrt.dynamo.lowering._partition import partition, get_submod_inputs
