+++ py/torch_tensorrt/dynamo/torch_compile/lowering/__init__.py
@@ -1,0 +1,7 @@
+from torch_tensorrt.dynamo.torch_compile.lowering._decompositions import (
+    get_decompositions,
+)
+from torch_tensorrt.dynamo.torch_compile.lowering._partition import (
+    partition,
+    get_submod_inputs,
+)
